Reorg tests and refactor server use
