+++ tests/e2e/test_spyre_basic.py
