+++ tests/e2e/test_spyre_embeddings.py
