+++ tests/e2e/test_spyre_seed.py
