+++ tests/worker/test_spyre_input_batch.py
